New frontend test file
